Fix #31

+++ mfc.sh
@@ -1,4 +1,4 @@
-#!/usr/bin/env bash
+#!/bin/bash
 
 
 MFC_PYTHON_MIN_MAJOR=3
@@ -6,7 +6,7 @@ MFC_PYTHON_MIN_MINOR=6
 
 # Check whether this script was called from MFC's root directory.
 if [ ! -f "$(pwd)/toolchain/mfc.py" ]; then
-    echo "[mfc.sh]: Error: You must call this script from within MFC's root folder."
+    echo "mfc: ERROR > You must call this script from within MFC's root folder."
 
     exit 1
 fi
@@ -16,8 +16,8 @@ source "$(pwd)/toolchain/util.sh"
 
 # Handle upgrading from older MFC build systems
 if [ -d "$(pwd)/bootstrap" ] || [ -d "$(pwd)/dependencies" ] || [ -f "$(pwd)/build/mfc.lock.yaml" ]; then
-    echo -e "$CYAN[mfc.sh]$COLOR_RESET:$RED Error: You are upgrading from an older version of MFC. Please remove, if applicable, the dependencies/, bootstrap/, and build/ directories before running this command again.$COLOR_RESET"
-
+    error "You are upgrading from an older version of$MAGENTA MFC$COLOR_RESET. Please remove, if applicable, the dependencies/, bootstrap/, and build/ directories before running this command again."
+    
     exit 1
 fi
 
@@ -37,16 +37,16 @@ if [ "$1" == "docker" ]; then
     shift;
 
     if ! command -v docker > /dev/null 2>&1; then
-        echo -e "$CYAN[mfc.sh]$COLOR_RESET:$RED Error: Docker is not installed.$COLOR_RESET"
-
+        error "$MAGENTA""Docker$COLOR_RESET is not installed."
+        
         exit 1
     fi
 
-    echo "$CYAN[mfc.sh]$COLOR_RESET: Running in Docker mode."
-    echo "  - Fetching image..."
+    log "Running in$MAGENTA Docker$COLOR_RESET mode."
+    log "  - Fetching image..."
     if ! docker pull henryleberre/mfc; then
-        echo -e "$CYAN[mfc.sh]$COLOR_RESET:$RED Error: Failed to fetch Docker image from Docker Hub.$COLOR_RESET"
-
+        error "Failed to fetch$MAGENTA Docker$COLOR_RESET image from$MAGENTA Docker Hub$COLOR_RESET."
+        
         exit 1
     fi
 
@@ -55,8 +55,8 @@ if [ "$1" == "docker" ]; then
                --mount type=bind,source="$(pwd)",target=/home/me/MFC \
                henryleberre/mfc
     if (($?)); then
-        echo -e "$CYAN[mfc.sh]$COLOR_RESET:$RED Error: Failed to start Docker container.$COLOR_RESET"
-
+        error "Failed to start Docker container."
+    
         exit 1
     fi
 
@@ -73,7 +73,7 @@ mkdir -p "$(pwd)/build"
 if [ -f "$(pwd)/build/cmake/bin/cmake" ]; then
     export PATH="$(pwd)/build/cmake/bin:$PATH"
 
-    echo -e "$CYAN[mfc.sh]$COLOR_RESET: Found CMake: $(pwd)/build/cmake/bin/cmake."
+    log "Found$MAGENTA CMake$COLOR_RESET: $MAGENTA$(pwd)/build/cmake/bin/cmake$COLOR_RESET."
 fi
 
 bShouldInstallCMake=false
@@ -81,7 +81,7 @@ if ! command -v cmake > /dev/null 2>&1; then
     # Not installed
     bShouldInstallCMake=true
 
-    echo -e "$CYAN[mfc.sh]$COLOR_RESET:$MAGENTA CMake$YELLOW is not installed.$COLOR_RESET"
+    warning "$MAGENTA""CMake$COLOR_RESET is not installed."
 else
     cmake_verstr=$(cmake --version | tr ' ' '\n' | sed -n 3p)
     cmake_major=$(echo $cmake_verstr | tr '.' '\n' | sed -n 1p)
@@ -99,54 +99,83 @@ else
         # Out of date
         bShouldInstallCMake=true
 
-        echo -e "$CYAN[mfc.sh]$COLOR_RESET:$MAGENTA CMake$YELLOW is out of date (current: $MAGENTA$cmake_verstr$YELLOW < minimum: $MAGENTA$MFC_CMAKE_MIN_VERSTR$YELLOW).$COLOR_RESET"
+        log "$MAGENTA CMake$YELLOW is out of date (current: $MAGENTA$cmake_verstr$YELLOW < minimum: $MAGENTA$MFC_CMAKE_MIN_VERSTR$YELLOW).$COLOR_RESET"
     fi
 fi
 
 if [ "$bShouldInstallCMake" = true ]; then
+    version="3.24.2"
+    arch="$(uname -m)"
+
+    bErrorDoSelfDownload=false
+    if [[ "$OSTYPE" == "darwin"* ]]; then
+        error "Cannot fetch$MAGENTA CMake$COLOR_RESET for$MAGENTA macOS$COLOR_RESET."
+        bErrorDoSelfDownload=true
+    fi
+
+    if ! [[ "$arch" -eq "x86_64" || "$arch" -eq "aarch64" ]]; then
+        error "Cannot fetch$MAGENTA CMake$COLOR_RESET for architecture $MAGENTA$arch$COLOR_RESET."
+        bErrorDoSelfDownload=true
+    fi
+
+    if [ "$bErrorDoSelfDownload" = true ]; then
+        log "Please install$MAGENTA CMake$COLOR_RESET manually:"
+        log " - via$MAGENTA Aptitude$COLOR_RESET:  $ sudo apt install cmake"
+        log " - via$MAGENTA Homebrew$COLOR_RESET:  $ brew install cmake"
+        log " - via$MAGENTA Spack$COLOR_RESET:     $ spack install cmake"
+        log " - via$MAGENTA Pacman$COLOR_RESET:    $ sudo pacman -S cmake"
+        log " - via$MAGENTA Lmod$COLOR_RESET:      $ module load cmake (if available)"
+        log " - via$MAGENTA cmake.org$COLOR_RESET: https://cmake.org/download/"
+
+        exit 1
+    fi
+
     if [ -d "$(pwd)/build/cmake" ]; then
         rm -rf "$(pwd)/build/cmake"
     fi
 
     mkdir -p "$(pwd)/build/cmake"
-
-    version="3.24.2"
-    arch="$(uname -m)"
+    
     filename="cmake-$version-linux-$arch.sh"
     repository="https://github.com/Kitware/CMake"
     url="$repository/releases/download/v$version/$filename"
 
-    echo -e "$CYAN[mfc.sh]$COLOR_RESET: Downloading$MAGENTA CMake v$version$COLOR_RESET for $MAGENTA$arch$COLOR_RESET from $CYAN$repository$COLOR_RESET."
+    log "Downloading$MAGENTA CMake v$version$COLOR_RESET for $MAGENTA$arch$COLOR_RESET from $CYAN$repository$COLOR_RESET."
 
-    if ! wget -P "$(pwd)/build/cmake" "$url"; then
-        echo -e "$CYAN[mfc.sh]$COLOR_RESET:$RED Error: Failed to download a compatible version of CMake."
-        echo "CMake is not discoverable or is an older release, incompatible with MFC. Please download"
-        echo "or install a recent version of CMake to get past this step. If you are currently on a"
-        echo "managed system like a cluster, provided there is no suitable environment module, you can"
-        echo "either build it from source, or get it via Spack."
-        echo "- The minimum required version is currently CMake v$MFC_CMAKE_MIN_MAJOR.$MFC_CMAKE_MIN_MINOR.$MFC_CMAKE_MIN_PATCH."
-        echo -e "- We attempted to download CMake v$version from $url.$COLOR_RESET"
+    cmake_fatal_error() {
+        log   "$MAGENTA""CMake$COLOR_RESET is not discoverable or is an older release, incompatible with$MAGENTA MFC$COLOR_RESET. Please download"
+        log   "or install a recent version of$MAGENTA CMake$COLOR_RESET to get past this step. If you are currently on a"
+        log   "managed system like a cluster, provided there is no suitable environment module, you can"
+        log   "either build it from source, or get it via$MAGENTA Spack$COLOR_RESET."
+        log   "- The minimum required version is currently$MAGENTA CMake$COLOR_RESET v$MFC_CMAKE_MIN_MAJOR.$MFC_CMAKE_MIN_MINOR.$MFC_CMAKE_MIN_PATCH."
+        log   "- We attempted to download$MAGENTA CMake$COLOR_RESET v$version from $url."
 
         exit 1
+    }
+
+    if ! command -v wget > /dev/null 2>&1; then
+        error "$MAGENTA""wget$COLOR_RESET is not installed but is necessary to download$MAGENTA CMake$COLOR_RESET."
+    
+        cmake_fatal_error
     fi
 
-    echo -e "$CYAN[mfc.sh]$COLOR_RESET: Installing$MAGENTA CMake$COLOR_RESET into $MAGENTA$(pwd)/build/cmake$COLOR_RESET."
+    if ! wget -P "$(pwd)/build/cmake" "$url"; then
+        error "Failed to download a compatible version of$MAGENTA CMake$COLOR_RESET."
+
+        cmake_fatal_error
+    fi
+
+    log "Installing$MAGENTA CMake$COLOR_RESET into $MAGENTA$(pwd)/build/cmake$COLOR_RESET."
 
     if ! $SHELL "$(pwd)/build/cmake/$filename" "--skip-license" "--prefix=$(pwd)/build/cmake"; then
-        echo -e "$CYAN[mfc.sh]$COLOR_RESET:$RED Error: Failed to install a compatible version of CMake."
-        echo "CMake is not discoverable or is an older release, incompatible with MFC. Please download"
-        echo "or install a recent version of CMake to get past this step. If you are currently on a"
-        echo "managed system like a cluster, provided there is no suitable environment module, you can"
-        echo "either build it from source, or get it via Spack."
-        echo "- The minimum required version is currently CMake v$MFC_CMAKE_MIN_MAJOR.$MFC_CMAKE_MIN_MINOR.$MFC_CMAKE_MIN_PATCH."
-        echo -e "- We attempted to install CMake v$version from $url.$COLOR_RESET"
-
-        exit 1
+        error "Failed to install a compatible version of CMake."
+        
+        cmake_fatal_error
     fi
 
     rm "$(pwd)/build/cmake/$filename"
 
-    echo -e "$CYAN[mfc.sh]$COLOR_RESET: Found$MAGENTA CMake$COLOR_RESET: $MAGENTA$(pwd)/build/cmake/bin/cmake$COLOR_RESET."
+    ok "Found$MAGENTA CMake$COLOR_RESET: $MAGENTA$(pwd)/build/cmake/bin/cmake$COLOR_RESET."
     export PATH="$(pwd)/build/cmake/bin:$PATH"
 fi
 
@@ -155,8 +184,8 @@ if [ -f "$(pwd)/build/venv/bin/activate" ]; then
     # Check Python is still working within the VENV
     if ! $(pwd)/build/venv/bin/python3 --version > /dev/null 2>&1; then
         # If not, delete it and install it again
-        echo -e "$CYAN[mfc.sh]$COLOR_RESET:$YELLOW WARNING: Python is no longer working inside the Virtualenv."
-        echo -e "                  Deleting the Virtualenv and starting from scratch... $COLOR_RESET"
+        warn "$MAGENTA""Python$COLOR_RESET is no longer working inside the Virtualenv."
+        warn "Deleting the Virtualenv and starting from scratch..."
 
         rm -r "$(pwd)/build/venv"
     fi
@@ -165,48 +194,50 @@ fi
 if ! command -v pip3 > /dev/null 2>&1 && [ ! -f "$(pwd)/build/venv/bin/activate" ]; then
     # Check whether python3 is in the $PATH / is accessible.
     if ! command -v python3 > /dev/null 2>&1; then
-        echo -e "$CYAN[mfc.sh]$COLOR_RESET:$RED Error: Couldn't find Python. Please ensure it is discoverable.$COLOR_RESET"
-
+        error "Couldn't find$MAGENTA Python$COLOR_RESET. Please ensure it is discoverable."
+    
         exit 1
     fi
 
     # CHeck Python's version for compatibility
     if ! python3 -c "import sys; exit(int(not (sys.version_info[0]==$MFC_PYTHON_MIN_MAJOR and sys.version_info[1] >= $MFC_PYTHON_MIN_MINOR)))"; then
-        echo -e "$CYAN[mfc.sh]$COLOR_RESET:$RED Error: $(python3 --version) is incompatible. Python v$MFC_PYTHON_MIN_MAJOR.$MFC_PYTHON_MIN_MINOR or higher is required.$COLOR_RESET"
-
+        error "$(python3 --version) is incompatible.$MAGENTA Python$COLOR_RESET v$MFC_PYTHON_MIN_MAJOR.$MFC_PYTHON_MIN_MINOR or higher is required."
+    
         exit 1
     fi
 
     get_pip_url="https://bootstrap.pypa.io/pip/get-pip.py"
 
-    echo -e "$CYAN[mfc.sh]$YELLOW Python's PIP is not installed.$COLOR_RESET"
-    echo -e "$CYAN[mfc.sh]$COLOR_RESET: Downloading Python's PIP from $get_pip_url..."
+    warning "$MAGENTA""Python$COLOR_RESET's$MAGENTA PIP$COLOR_RESET is not installed."
+    log     "Downloading$MAGENTA Python$COLOR_RESET's$MAGENTA PIP$COLOR_RESET from $get_pip_url..."
 
     if ! wget -O "$(pwd)/build/get-pip.py" "$get_pip_url"; then
-        echo -e "$CYAN[mfc.sh]$COLOR_RESET:$RED Error: Couldn't download get-pip.py.$COLOR_RESET"
-
+        error "Couldn't download get-pip.py."
+    
         exit 1
     fi
 
     # Suppress PIP version warning (out of date)
     export PIP_DISABLE_PIP_VERSION_CHECK=1
     if ! python3 "$(pwd)/build/get-pip.py" --user; then
-        echo -e "$CYAN[mfc.sh]$COLOR_RESET:$RED Error: Couldn't install pip with get-pip.py.$COLOR_RESET"
-
+        error "Couldn't install$MAGENTA pip$COLOR_RESET with get-pip.py"
+    
         exit 1
     fi
+
+    ok "Installed pip."
 fi
 
 
 # Create a Python virtualenv if it hasn't already been created
 if [ ! -f "$(pwd)/build/venv/bin/activate" ]; then
-    echo -e "$CYAN[mfc.sh]$COLOR_RESET: Creating a Python virtual environment (venv)..."
-
     if ! python3 -m venv "$(pwd)/build/venv"; then
-        echo -e "$CYAN[mfc.sh]$COLOR_RESET:$RED Error: Failed to create a Python virtual environment. Delete the build/venv folder and try again.$COLOR_RESET"
-
+        error "Failed to create a$MAGENTA Python$COLOR_RESET virtual environment. Delete the build/venv folder and try again."
+    
         exit 1
     fi
+
+    ok "Created a$MAGENTA Python$COLOR_RESET virtual environment (venv)."
 fi
 
 
@@ -216,19 +247,19 @@ if [ "$1" == "venv" ]; then
     if [[ "$VIRTUAL_ENV" != "" ]]; then
         # Already inside the venv, exit out of it
 
-        echo " > Exiting the MFC Python virtual environment."
+        log " > Exiting the$MAGENTA MFC Python$COLOR_RESET virtual environment."
 
         deactivate
     else
         # Enter the venv
-        echo " > Entering the MFC Python virtual environment."
+        log " > Entering the$MAGENTA MFC Python$COLOR_RESET virtual environment."
 
         source "$(pwd)/build/venv/bin/activate"
 
-        echo " > To exit, you can do any of the following:"
-        echo "    - Run 'deactivate'."
-        echo "    - Run '. ./mfc.sh venv'."
-        echo "    - Close your terminal."
+        log " > To exit, you can do any of the following:"
+        log "    - Run 'deactivate'."
+        log "    - Run '. ./mfc.sh venv'."
+        log "    - Close your terminal."
     fi
 
     return
@@ -236,8 +267,8 @@ fi
 
 
 # Activate the Python venv
-echo -e "$CYAN[mfc.sh]$COLOR_RESET: Entering the Python virtual environment (venv)."
 source "$(pwd)/build/venv/bin/activate"
+ok "Entered the$MAGENTA Python$COLOR_RESET virtual environment (venv)."
 
 
 # Fetch required Python modules.
@@ -250,13 +281,13 @@ for module in "${REQUIRED_PYTHON_MODULES[@]}"; do
     install_name=$(echo $module | tr ',' '\n' | tail -n 1)
 
     if ! python3 -c "import $import_name" > /dev/null 2>&1; then
-        echo -e "$CYAN[mfc.sh]$COLOR_RESET: Installing Python PIP package $MAGENTA$install_name$COLOR_RESET (into venv)."
-
-        if ! pip3 install "$install_name"; then
-            echo -e "$CYAN[mfc.sh]$COLOR_RESET:$RED Error: Failed to install $import_name/$install_name through Python3's pip.$COLOR_RESET"
-
+        if ! PIP_DISABLE_PIP_VERSION_CHECK=1 pip3 install "$install_name"; then
+            error "Failed to install $import_name/$install_name through Python3's pip."
+        
             exit 1
         fi
+
+        ok "Installed$MAGENTA Python PIP$COLOR_RESET package $MAGENTA$install_name$COLOR_RESET (into venv)."
     fi
 done
 
@@ -267,7 +298,7 @@ code=$?
 
 
 # Deactivate the Python virtualenv in case the user "source"'d this script
-echo -e "$CYAN[mfc.sh]$COLOR_RESET: Exiting the Python virtual environment."
+log "Exiting the$MAGENTA Python$COLOR_RESET virtual environment."
 deactivate
